+++ opengl_engine/Maths/Quaternion.cpp
@@ -64,17 +64,17 @@ Vector3 Quaternion::toEuler(const Quaternion& quat)
 	//  roll (x-axis rotation)
 	double sinr_cosp = 2.0 * (quat.w * quat.x + quat.y * quat.z);
 	double cosr_cosp = 1.0 - 2.0 * (quat.x * quat.x + quat.y * quat.y);
-	roll = Maths::atan2(sinr_cosp, cosr_cosp);
+	roll = Maths::atan2(static_cast<float>(sinr_cosp), static_cast<float>(cosr_cosp));
 
 	//  pitch (y-axis rotation)
 	double sinp = 2.0 * (quat.w * quat.y - quat.z * quat.x);
-	if (std::abs(sinp) >= 1) pitch = Maths::copysign(Maths::pi / 2.0f, sinp); //  use 90 degrees if out of range
-	else pitch = std::asin(sinp);
+	if (std::abs(sinp) >= 1) pitch = Maths::copysign(Maths::pi / 2.0f, static_cast<float>(sinp)); //  use 90 degrees if out of range
+	else pitch = static_cast<float>(std::asin(sinp));
 
 	//  yaw (z-axis rotation)
 	double siny_cosp = 2.0 * (quat.w * quat.z + quat.x * quat.y);
 	double cosy_cosp = 1.0 - 2.0 * (quat.y * quat.y + quat.z * quat.z);
-	yaw = Maths::atan2(siny_cosp, cosy_cosp);
+	yaw = Maths::atan2(static_cast<float>(siny_cosp), static_cast<float>(cosy_cosp));
 
 	return Vector3{ yaw, roll, pitch };
 }
